Added basic error handling
This change only stops the user from saving data with null values(no given amount, category, etc.) and makes it so read/writeJson function stop if an error occurs when working with files.
Errors are not documented in any way as we do not have a log system.
And currently the user is not informed of errors.