kinda fixed the error on loading transactions too quickly
The error came from the transactions list not being initialized. But the true problem of the transactions not being displayed was only solved by moving the reading from database function back to the transactions page.
Which means that transactions will only be loaded if that page is loaded first.
If the reading is called from a shared razor file like MainLayout, then I could not find a way for the transactions page to update automatically after the database was read.
This also means that data is fetched from the database each time the page is loaded, which is not ideal.